+++ changelog.md
@@ -8,6 +8,7 @@
 * **new** `MaxMind GeoLite2` support for weekly downloading of Lite database
 * **new** option to hide the `Object` column from the Logs
 * **new** plugin dashboard widget for the GEO Location information
+* **new** component `DebugPress` expanded with two new events
 * **new** logs panel option to filter by country based on geolocation
 * **new** logs panel popup dialog with overview of all event data split in tabs
 * **edit** changes in the order for some columns on the log panel